✨ feat: create a deploy button and move depoy event handler to deploy button
