+++ src/lib/hooks/useDeployEventHandler.ts
@@ -1,29 +1,20 @@
 import { ChangeEvent } from "react";
-import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
-import { setCookie } from "cookies-next";
+import { useSetRecoilState } from "recoil";
 import { SelectChangeEvent } from "@mui/material";
 
-import { getOrgRepos, getUserRepos, deployRepo } from "../api";
-import { gitRepoState, cloneUrlState, cloneRepoName } from "../recoil/git";
-import buildOptionsState from "../recoil/userBuildOptions";
-import userDeploymentsState from "../recoil/userDeployments";
-import getFormattedKoreaTime from "../../lib/utils/getFormattedKoreaTime";
+import { getOrgRepos, getUserRepos } from "lib/api";
+import { gitRepoState } from "lib/recoil/git";
+import buildOptionsState from "lib/recoil/userBuildOptions";
 
-import { UserDeploymentData } from "../../types/deployment";
 import {
   Repo,
   BuildOptions,
   EventHandlerName,
-} from "../../types/projectOption";
+} from "types/projectOption";
 
 function useDeployEventHandler(type: EventHandlerName, userId?: string) {
-  const [repoCloneUrl, setCloneUrl] = useRecoilState<string>(cloneUrlState);
-  const [buildOption, setBuildOption] =
-    useRecoilState<BuildOptions>(buildOptionsState);
-  const [deploymentList, setDeploymentList] =
-    useRecoilState<UserDeploymentData[]>(userDeploymentsState);
+  const setBuildOption = useSetRecoilState<BuildOptions>(buildOptionsState);
   const setGitRepos = useSetRecoilState<Repo[]>(gitRepoState);
-  const repoName = useRecoilValue<string>(cloneRepoName);
 
   switch (type) {
     case "spaceChange": {
@@ -31,7 +22,6 @@ function useDeployEventHandler(type: EventHandlerName, userId?: string) {
         if (!userId) return;
 
         setBuildOption(prev => ({ ...prev, userId }));
-
         const selectedNamespaceUrl = e.target.value;
 
         const spaceOption = selectedNamespaceUrl
@@ -48,20 +38,16 @@ function useDeployEventHandler(type: EventHandlerName, userId?: string) {
             selectedNamespace,
           );
 
-          return setGitRepos(orgRepos);
+          setGitRepos(orgRepos);
+
+          return;
         }
 
         const { data: userRepos } = await getUserRepos(userId);
 
-        return setGitRepos(userRepos);
-      };
-    }
-    case "repoChange": {
-      return (e: SelectChangeEvent) => {
-        const selectedRepoUrl = e.target.value;
+        setGitRepos(userRepos);
 
-        setCloneUrl(selectedRepoUrl);
-        setBuildOption(prev => ({ ...prev, repoCloneUrl: selectedRepoUrl }));
+        return;
       };
     }
     case "nodeVersionChange": {
@@ -101,6 +87,7 @@ function useDeployEventHandler(type: EventHandlerName, userId?: string) {
         }));
       };
     }
+
     case "removeEnvClick": {
       return (curIndex: number) => {
         setBuildOption(prev => ({
@@ -109,41 +96,7 @@ function useDeployEventHandler(type: EventHandlerName, userId?: string) {
         }));
       };
     }
-    case "deployClick": {
-      return async () => {
-        if (!userId) return;
 
-        const filteredEnvs = buildOption.envList.filter((_, i) => i !== 0);
-        const formattedTime = getFormattedKoreaTime(new Date());
-
-        const userBuildOptions = {
-          userId,
-          repoName,
-          repoCloneUrl,
-          repoUpdatedAt: formattedTime,
-          nodeVersion: buildOption.nodeVersion,
-          installCommand: buildOption.installCommand,
-          buildCommand: buildOption.buildCommand,
-          envList: filteredEnvs,
-          buildType: buildOption.buildType,
-          lastCommitMessage: "",
-        };
-
-        const { data: userDeploymentData } = await deployRepo(userBuildOptions);
-
-        const copyUserDeployData = JSON.parse(
-          JSON.stringify(userDeploymentData),
-        );
-        copyUserDeployData.buildingLog = [];
-        setDeploymentList([...deploymentList, userDeploymentData]);
-        setCookie(
-          "userDeployments",
-          JSON.stringify([...deploymentList, copyUserDeployData]),
-        );
-
-        return userDeploymentData;
-      };
-    }
     default:
       return () => {};
   }
